refactor: add some folders and move some styles
