+++ rsteam/apps/rslang/src/components/UI/navbar/Navbar.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import {Link} from 'react-router-dom';
-import './Navbar.scss';
 import Burger from '../burger/Burger';
 
 const Navbar = () => {
