+++ rsteam/apps/rslang/src/components/UI/rsschool/RSSchoolImg.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import './RSSchoolImg.scss';
 
 const RSSchoolImg = () => {
   return (
