+++ rsteam/apps/rslang/src/main.scss
@@ -45,3 +45,7 @@ $logo-font: 'Source Serif Pro', serif;
   margin: 0 auto;
   padding: 10px;
 }
+
+.main {
+  flex: 1 0 auto;
+}
