--- rsteam/apps/rslang/src/pages/Login.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const Login = () => {
-  return (
-    <div>
-      
-    </div>
-  );
-};
-
-export default Login;
