--- rsteam/apps/rslang/src/pages/Ourteam.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const Ourteam = () => {
-  return (
-    <div>
-      <h1>Our team page</h1>
-    </div>
-  );
-};
-
-export default Ourteam;
