--- rsteam/apps/rslang/src/pages/Page404.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const Page404 = () => {
-  return (
-    <div>
-      <h1>There is no such page</h1>
-    </div>
-  );
-};
-
-export default Page404;
