--- rsteam/apps/rslang/src/pages/SingUp.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const SingUp = () => {
-  return (
-    <div>
-      
-    </div>
-  );
-};
-
-export default SingUp;
