--- rsteam/apps/rslang/src/pages/Statistics.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const Statistics = () => {
-  return (
-    <div>
-      <h1>Statistics page</h1>
-    </div>
-  );
-};
-
-export default Statistics;
