--- rsteam/apps/rslang/src/pages/Textbook.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const Textbook = () => {
-  return (
-    <div>
-      <h1>Textbook page</h1>
-    </div>
-  );
-};
-
-export default Textbook;
