+++ rsteam/apps/rslang/src/pages/home/Home.scss
@@ -1,5 +1,4 @@
 .home {
-  flex: 1 0 auto;
   background-color: #ffffff;
-  box-shadow: inset 0 3px 3px rgba(0, 0, 0, 0.3);
+  box-shadow: inset 3px 3px 3px rgba(0, 0, 0, 0.3);
 }
