+++ rsteam/apps/rslang/src/pages/page404/Page404.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import Footer from '../../components/UI/footer/Footer';
 
-const Home = () => {
+const Page404 = () => {
   return (
     <div className='container-wrapper'>
-      <main className='home main'>
+      <main className='page404 main'>
         <div className='container-inner'>
-          <h1>RsLang</h1>
+          <h1>There is no such page</h1>
         </div>
       </main>
       <Footer />
@@ -14,4 +14,4 @@ const Home = () => {
   );
 };
 
-export default Home;
+export default Page404;
