+++ rsteam/apps/rslang/src/pages/statistics/Statistics.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import Footer from '../../components/UI/footer/Footer';
 
-const Home = () => {
+const Statistics = () => {
   return (
     <div className='container-wrapper'>
       <main className='home main'>
         <div className='container-inner'>
-          <h1>RsLang</h1>
+          <h1>Statistics page</h1>
         </div>
       </main>
       <Footer />
@@ -14,4 +14,4 @@ const Home = () => {
   );
 };
 
-export default Home;
+export default Statistics;
