+++ rsteam/apps/rslang/src/router/index.ts
@@ -1,11 +1,11 @@
 import Home from '../pages/home/Home';
-import Ourteam from '../pages/Ourteam';
-import Textbook from '../pages/Textbook';
+import Ourteam from '../pages/our-team/Ourteam';
+import Textbook from '../pages/textbook/Textbook';
 import Savannah from '../pages/Savannah';
 import Audiocall from '../pages/Audiocall';
-import Statistics from '../pages/Statistics';
-import Login from '../pages/Login';
-import SignUp from '../pages/SingUp';
+import Statistics from '../pages/statistics/Statistics';
+import Login from '../pages/login/Login';
+import SignUp from '../pages/signup/SingUp';
 
 export const routes = [
   {path: "/", element: Home},
