+++ rsteam/apps/rslang/src/styles.scss
@@ -1,1 +1,6 @@
-@import 'main';
+@import './main.scss';
+@import './pages/home/Home.scss';
+@import './components/UI/burger/Burger.scss';
+@import './components/UI/footer/Footer.scss';
+@import './components/UI/navbar/Navbar.scss';
+@import './components/UI/rsschool/RSSchoolImg.scss';
